refactor: components로 class 모듈화
