+++ src/index.js
@@ -1,5 +1,5 @@
-import RacingGame from './RacingGame.js';
-import init from './init.js';
+import RacingGame from './components/RacingGame.js';
+import init from './components/init.js';
 
 init();
 new RacingGame();
